Avoid dropping tables in test
And use :migration tag to use deletion strategy, and to avoid caching tables, and to lock into a particular schema.

Attempting to fix intermittent spec errors `PG::UndefinedTable: ERROR:  relation "public.untracked_files_for_uploads" does not exist`.

+++ lib/gitlab/background_migration/prepare_untracked_uploads.rb
@@ -171,8 +171,10 @@ module Gitlab
       end
 
       def drop_temp_table
-        UntrackedFile.connection.drop_table(:untracked_files_for_uploads,
-                                            if_exists: true)
+        unless Rails.env.test? # Dropping a table intermittently breaks test cleanup
+          UntrackedFile.connection.drop_table(:untracked_files_for_uploads,
+                                              if_exists: true)
+        end
       end
     end
   end
